getIdentityToken {rstudioapi}	R Documentation

Get the User's Identity Token

Description

Retrieve the user's OpenID Connect identity token for the current Posit Workbench session, if any. This token can be used to authenticate with services that trust Workbench's identity provider.

Usage

getIdentityToken()

Value

A list containing:

token

The identity token string.

expiry

The token expiry time as a POSIXct datetime object.

Throws an error if the token cannot be retrieved.

Note

This function works for any IDE running within a Posit Workbench session, not just RStudio.

[Package rstudioapi version 0.18.0 Index]