getOAuthCredentials {rstudioapi}	R Documentation

Retrieve OAuth Credentials for Integrations

Description

Retrieve OAuth credentials for a configured OAuth integration in Posit Workbench. This function exchanges the current session for OAuth credentials that can be used to authenticate with external services. This works in any IDE running within a Posit Workbench session.

Usage

getOAuthCredentials(audience)

Arguments

audience
The GUID of the OAuth integration configured in Posit Workbench.

Value

A list containing:

access_token

The OAuth access token.

expiry

The token expiry time as a POSIXct datetime object.

audience

The integration GUID (audience) that was used to retrieve the credentials.

Throws an error if the credentials cannot be retrieved or the integration is not found.

Note

This function requires Posit Workbench version 2026.01.0 or later. It works in any IDE running within a Posit Workbench session (not just RStudio).

Examples

## Not run: 
# Retrieve OAuth credentials for an integration
creds <- getOAuthCredentials("4c1cfecb-1927-4f19-bc2f-d8ac261364e0")
if (!is.null(creds)) {
  cat("Access token:", creds$access_token, "\n")
  cat("Expires at:", format(creds$expiry), "\n")
}

## End(Not run)

[Package rstudioapi version 0.18.0 Index]