getThemeInfo {rstudioapi}	R Documentation

Retrieve Themes

Description

Retrieves a list with themes information. Currently, editor as the theme used under the code editor, global as the global theme applied to the main user interface in RStudio and dark when the user interface is optimized for dark themes.

Usage

getThemeInfo()

Details

RStudio 1.2.553 adds support for background and color as the primary colors used under the code editor.

[Package rstudioapi version 0.8 Index]