getVersion {rstudioapi}	R Documentation

Determine the version of RStudio

Description

Use getVersion() to determine the current version of RStudio. This can be useful for R packages which need to gate certain functionality based on the version of RStudio currently available.

Usage

getVersion()

Value

A "numeric_version" object, giving the version of RStudio in use.

[Package rstudioapi version 0.18.0 Index]