hasFun {rstudioapi}	R Documentation

Exists/get for RStudio functions

Description

These are specialized versions of get and exists that look in the rstudio package namespace. If RStudio is not running, hasFun will return FALSE.

Usage

hasFun(name, version_needed = NULL, ...)

findFun(name, version_needed = NULL, ...)

Arguments

name
name of object to look for

version_needed
An optional version specification. If supplied, ensures that RStudio is at least that version. This is useful if function behavior has changed over time.

...
other arguments passed on to exists and get

Examples

rstudioapi::hasFun("viewer")

[Package rstudioapi version 0.15.0 Index]