Check if RStudio is running.

Check if RStudio is running.

isAvailable(version_needed = NULL, child_ok = FALSE)

verifyAvailable(version_needed = NULL)

Arguments

version_needed
An optional version specification. If supplied, ensures that RStudio is at least that version.

child_ok
Boolean; check if the current R process is a child process of the main RStudio session? This can be useful for e.g. RStudio Jobs, where you'd like to communicate back with the main R session from a child process through rstudioapi.

Value

isAvailable a boolean; verifyAvailable an error message if RStudio is not running

Examples

rstudioapi::isAvailable()

#> [1] FALSE

if (FALSE) rstudioapi::verifyAvailable()