jobRunScript {rstudioapi}	R Documentation

Run R Script As Background Job

Description

Starts an R script as a background job.

Usage

jobRunScript(
  path,
  name = NULL,
  encoding = "unknown",
  workingDir = NULL,
  importEnv = FALSE,
  exportEnv = ""
)

Arguments

path
The path to the R script to be run.

name
A name for the background job. When NULL (the default), the filename of the script is used as the job name.

encoding
The text encoding of the script, if known.

workingDir
The working directory in which to run the job. When NULL (the default), the parent directory of the R script is used.

importEnv
Whether to import the global environment into the job.

exportEnv
The name of the environment in which to export the R objects created by the job. Use "" (the default) to skip export, "R_GlobalEnv"' to export to the global environment, or the name of an environment object to create an object with that name.

See Also

Other jobs: jobAddOutput(), jobAddProgress(), jobAdd(), jobRemove(), jobSetProgress(), jobSetState(), jobSetStatus()

[Package rstudioapi version 0.15.0 Index]