launcher {rstudioapi}	R Documentation

Retrieve Launcher Information

Description

Retrieve information about the launcher, as well as the different clusters that the launcher has been configured to use.

Check if the RStudio launcher is available and configured to support 'ad-hoc' jobs; that is, jobs normally launched by the user through the RStudio IDE's user interface.

Retrieve information on launcher jobs.

Usage

launcherGetInfo()

launcherAvailable()

launcherGetJobs(statuses = NULL, fields = NULL, tags = NULL,
  includeSessions = FALSE)

Arguments

statuses
Return only jobs whose status matches one of statuses. Valid statuses are: Pending, Running, Suspended, Failed, Finished, Killed, Canceled. When NULL, all jobs are returned.

fields
Return a subset of fields associated with each job object. When NULL, all fields associated with a particular job are returned.

tags
An optional set of tags. Only jobs that have been assigned one of these requested tags will be returned.

includeSessions
Boolean; include jobs which are also operating as RStudio R sessions?

[Package rstudioapi version 0.10 Index]