launcherAvailable {rstudioapi}	R Documentation

Check if Workbench Launcher is Available

Description

Check if the Workbench launcher is available and configured to support Workbench jobs; that is, jobs normally launched by the user through the RStudio IDE's user interface.

Usage


launcherAvailable()

See Also

Other job-launcher functionality: launcherConfig(), launcherContainer(), launcherControlJob(), launcherGetInfo(), launcherGetJob(), launcherGetJobs(), launcherHostMount(), launcherNfsMount(), launcherPlacementConstraint(), launcherResourceLimit(), launcherSubmitJob(), launcherSubmitR()

[Package rstudioapi version 0.16.0 Index]