launcherContainer {rstudioapi}	R Documentation

Define a Workbench Launcher Container

Description

Define a launcher container, suitable for use with the container argument to launcherSubmitJob().

Usage


launcherContainer(image, runAsUserId = NULL, runAsGroupId = NULL)

Arguments

image
The container image to use.

runAsUserId
The user id to run as within the container. Defaults to the container-specified user.

runAsGroupId
The group id to run as within the container. Defaults to the container-specified group.

See Also

Other job-launcher functionality: launcherAvailable(), launcherConfig(), launcherControlJob(), launcherGetInfo(), launcherGetJob(), launcherGetJobs(), launcherHostMount(), launcherNfsMount(), launcherPlacementConstraint(), launcherResourceLimit(), launcherSubmitJob(), launcherSubmitR()

[Package rstudioapi version 0.16.0 Index]