launcherControlJob {rstudioapi}	R Documentation

Interact with (Control) a Workbench Job

Description

Interact with a Workbench job.

Usage


launcherControlJob(
  jobId,
  operation = c("suspend", "resume", "stop", "kill", "cancel")
)

Arguments

jobId
The job id.

operation
The operation to execute. The operation should be one of c("suspend", "resume", "stop", "kill", "cancel"). Note that different launcher plugins support different subsets of these operations – consult your launcher plugin documentation to see which operations are supported.

See Also

Other job-launcher functionality: launcherAvailable(), launcherConfig(), launcherContainer(), launcherGetInfo(), launcherGetJob(), launcherGetJobs(), launcherHostMount(), launcherNfsMount(), launcherPlacementConstraint(), launcherResourceLimit(), launcherSubmitJob(), launcherSubmitR()

[Package rstudioapi version 0.16.0 Index]