launcherHostMount {rstudioapi}	R Documentation

Define a Workbench Launcher Host Mount

Description

Define a launcher host mount, suitable for use with the mounts argument to launcherSubmitJob(). This can be used to mount a path from the host into the generated container.

Usage

launcherHostMount(path, mountPath, readOnly = TRUE)

Arguments

path
The host path to be mounted.

mountPath
The destination path for the mount in the container.

readOnly
Boolean; should the path be mounted read-only?

See Also

Other job-launcher functionality: launcherAvailable(), launcherConfig(), launcherContainer(), launcherControlJob(), launcherGetInfo(), launcherGetJobs(), launcherGetJob(), launcherNfsMount(), launcherPlacementConstraint(), launcherResourceLimit(), launcherSubmitJob(), launcherSubmitR()

[Package rstudioapi version 0.15.0 Index]