launcherNfsMount {rstudioapi}	R Documentation

Define a Workbench Launcher NFS Mount

Description

Define a launcher NFS mount, suitable for use with the mounts argument to launcherSubmitJob(). This can be used to mount a path from a networked filesystem into a newly generated container.

Usage

launcherNfsMount(host, path, mountPath, readOnly = TRUE)

Arguments

host
The host name, or IP address, of the NFS server.

path
The NFS path to be mounted.

mountPath
The destination path for the mount in the container.

readOnly
Boolean; should the path be mounted read-only?

See Also

Other job-launcher functionality: launcherAvailable(), launcherConfig(), launcherContainer(), launcherControlJob(), launcherGetInfo(), launcherGetJobs(), launcherGetJob(), launcherHostMount(), launcherPlacementConstraint(), launcherResourceLimit(), launcherSubmitJob(), launcherSubmitR()

[Package rstudioapi version 0.15.0 Index]