launcherPlacementConstraint {rstudioapi}	R Documentation

Define a Workbench Launcher Placement Constraint

Description

Define a launcher placement constraint, suitable for use with the placementConstraints argument to launcherSubmitJob().

Usage


launcherPlacementConstraint(name, value = NULL)

Arguments

name
The name of this placement constraint.

value
The value of the constraint. A job will only be placed on a requested node if the requested placement constraint is present.

See Also

Other job-launcher functionality: launcherAvailable(), launcherConfig(), launcherContainer(), launcherControlJob(), launcherGetInfo(), launcherGetJob(), launcherGetJobs(), launcherHostMount(), launcherNfsMount(), launcherResourceLimit(), launcherSubmitJob(), launcherSubmitR()

[Package rstudioapi version 0.16.0 Index]