launcherResourceLimit {rstudioapi}	R Documentation

Define a Workbench Launcher Resource Limit

Description

Define a launcher resource limit, suitable for use with the resourceLimits argument to launcherSubmitJob().

Usage


launcherResourceLimit(type, value)

Arguments

type
The resource limit type. Must be one of cpuCount, cpuFrequency, cpuSet, cpuTime, memory, memorySwap. Different launcher plugins may support different subsets of these resource limit types; please consult the plugin documentation to learn which limits are supported.

value
The formatted value of the requested limit.

See Also

Other job-launcher functionality: launcherAvailable(), launcherConfig(), launcherContainer(), launcherControlJob(), launcherGetInfo(), launcherGetJob(), launcherGetJobs(), launcherHostMount(), launcherNfsMount(), launcherPlacementConstraint(), launcherSubmitJob(), launcherSubmitR()

[Package rstudioapi version 0.16.0 Index]