primary_selection {rstudioapi}	R Documentation

Extract the Primary Selection

Description

By default, functions returning a document context will return a list of selections, including both the 'primary' selection and also 'other' selections (e.g. to handle the case where a user might have multiple cursors active). Use primary_selection() to extract the primary selection.

Usage


primary_selection(x, ...)

Arguments

x
A document context, or a selection.

...
Optional arguments (currently ignored).

[Package rstudioapi version 0.16.0 Index]