readPreference {rstudioapi}	R Documentation

Read Preference

Description

Reads a user preference, useful to remember preferences across different R sessions for the same user.

Usage

readPreference(name, default)

Arguments

name
The name of the preference.

default
The default value to use when the preference is not available.

Details

User preferences can have arbitrary names and values. You must write the preference with writePreference before it can be read (otherwise its default value will be returned).

Note

The readPreference function was added in version 1.1.67 of RStudio.

See Also

readRStudioPreference, which reads RStudio IDE preferences.

[Package rstudioapi version 0.15.0 Index]