restartSession {rstudioapi}	R Documentation

Restart the R Session

Description

Restart the RStudio session.

Usage


restartSession(command = "", clean = FALSE)

Arguments

command
A command (as a string) to be run after restarting.

clean
Boolean; when FALSE, the current R session (including loaded packages and data objects) will be saved and restored in the new session.

Note

The restartSession function was added in version 1.1.281 of RStudio. Support for the clean argument was added for version 2024.04 release of RStudio; it is silently ignored in older versions of RStudio.

[Package rstudioapi version 0.16.0 Index]