setGhostText {rstudioapi}	R Documentation

Set ghost text

Description

Set ghost text in the current document. The ghost text will be inserted at the current cursor position. Ghost text can be inserted into the document by pressing Tab, and will be automatically dismissed if the user navigates the cursor away.

Usage


setGhostText(text)

Arguments

text
The ghost text to set.

[Package rstudioapi version 0.16.0 Index]