showDialog {rstudioapi}	R Documentation

Show Dialog Box

Description

Shows a dialog box with a given title and contents.

Usage

showDialog(title, message, url = "")

Arguments

title
The title to display in the dialog box.

message
A character vector with the contents to display in the main dialog area. Contents can contain the following HTML tags: "p", "em", "strong", "b" and "i".

url
An optional url to display under the message.

Details

 showDialog("A dialog", "Showing <b>bold</b> text in the
message.") 

Note

The showDialog function was added in version 1.1.67 of RStudio.

[Package rstudioapi version 0.15.0 Index]