showPrompt {rstudioapi}	R Documentation

Show Prompt Dialog Box

Description

Shows a dialog box with a prompt field.

Usage


showPrompt(title, message, default = NULL)

Arguments

title
The title to display in the dialog box.

message
A character vector with the contents to display in the main dialog area.

default
An optional character vector that fills the prompt field with a default value.

Note

The showPrompt function was added in version 1.1.67 of RStudio.

[Package rstudioapi version 0.16.0 Index]