showQuestion {rstudioapi}	R Documentation

Show Question Dialog Box

Description

Shows a dialog box asking a question.

Usage

showQuestion(title, message, ok = NULL, cancel = NULL)

Arguments

title
The title to display in the dialog box.

message
A character vector with the contents to display in the main dialog area.

ok
And optional character vector that overrides the caption for the OK button.

cancel
An optional character vector that overrides the caption for the Cancel button.

Note

The showQuestion function was added in version 1.1.67 of RStudio.

[Package rstudioapi version 0.15.0 Index]