terminalCreate {rstudioapi}	R Documentation

Create a Terminal

Description

Create a new Terminal.

Usage

terminalCreate(caption = NULL, show = TRUE)

Arguments

caption
The desired terminal caption. When NULL or blank, the terminal caption will be chosen by the system.

show
If FALSE, terminal won't be brought to front.

Value

The terminal identifier as a character vector (NULL if unable to create the terminal or the given terminal caption is already in use).

Note

The terminalCreate function was added in version 1.1.350 of RStudio.

Examples

## Not run: 
termId <- rstudioapi::terminalCreate('My Terminal')

## End(Not run)

[Package rstudioapi version 0.7.0-9000 Index]