updateDialog {rstudioapi}	R Documentation

Updates a Dialog Box

Description

Updates specific properties from the current dialog box.

Usage

updateDialog(...)

Arguments

...
Named parameters and values to update a dialog box.

Details

Currently, the only dialog with support for this action is the New Connection dialog in which the code preview can be updated through this API.

 updateDialog(code = "con <- NULL") 

Note

The updateDialog function was added in version 1.1.67 of RStudio.

[Package rstudioapi version 0.15.0 Index]