writePreference {rstudioapi}	R Documentation

Write Preference

Description

Writes a user interface preference, useful to remember preferences across different r sessions for the same user.

Usage

writePreference(name, value)

Arguments

name
The name of the preference.

value
The value of the preference.

Note

The writePreference function was added in version 1.1.67 of RStudio.

[Package rstudioapi version 0.10 Index]