askForPassword {rstudioapi}	R Documentation

Ask the user for a password interactively

Description

Ask the user for a password interactively.

Usage

askForPassword(prompt)

Arguments

prompt
Single element character vector containing the prompt to be displayed

Details

RStudio also sets the global askpass option to the rstudioapi::askForPassword function so that it can be invoked in a front-end independent manner.

Note

The askForPassword function was added in version 0.99.853 of RStudio.

Examples

## Not run: 
rstudioapi::askForPassword("Please enter your password")

## End(Not run)

[Package rstudioapi version 0.10 Index]