askForSecret {rstudioapi}	R Documentation

Prompt user for secret

Description

Request a secret from the user. If the keyring package is installed, it will be used to cache requested secrets.

Usage

askForSecret(
  name,
  message = paste(name, ":", sep = ""),
  title = paste(name, "Secret")
)

Arguments

name
The name of the secret.

message
A character vector with the contents to display in the main dialog area.

title
The title to display in the dialog box.

Note

The askForSecret function was added in version 1.1.419 of RStudio.

[Package rstudioapi version 0.13 Index]