askForSecret {rstudioapi}	R Documentation

Show Prompt for Secret Dialog

Description

Shows a dialog box asking for a secret with support to remember such secret using the 'keyring' package.

Usage

askForSecret(name, message = paste(name, ":", sep = ""),
  title = paste(name, "Secret"))

Arguments

name
The name of the secret.

message
A character vector with the contents to display in the main dialog area.

title
The title to display in the dialog box.

Note

The askForSecret function was added in version 1.1.419 of RStudio.

[Package rstudioapi version 0.10 Index]