bugReport {rstudioapi}	R Documentation

File an RStudio Bug Report

Description

A utility function to assist with the filing of an RStudio bug report. This function will pre-populate a template with information useful in understanding your reported bug.

Usage


bugReport()

[Package rstudioapi version 0.16.0 Index]