callFun {rstudioapi}	R Documentation

Call an RStudio API function

Description

This function will return an error if RStudio is not running, or the function is not available. If you want to fall back to different behavior, use hasFun.

Usage

callFun(fname, ...)

Arguments

fname
name of the RStudio function to call.

...
Other arguments passed on to the function

Examples

if (rstudioapi::isAvailable()) {
  rstudioapi::callFun("versionInfo")
}

[Package rstudioapi version 0.10 Index]