Call an RStudio API function

This function will return an error if RStudio is not running, or the function is not available. If you want to fall back to different behavior, use hasFun.

callFun(fname, ...)

Arguments

fname
name of the RStudio function to call.

...
Other arguments passed on to the function

Examples

if (rstudioapi::isAvailable()) {
  rstudioapi::callFun("versionInfo")
}

#> $citation
#> 
#> To cite RStudio in publications use:
#> 
#>   RStudio Team (2016). RStudio: Integrated Development for R. RStudio,
#>   Inc., Boston, MA URL http://www.rstudio.com/.
#> 
#> A BibTeX entry for LaTeX users is
#> 
#>   @Manual{,
#>     title = {RStudio: Integrated Development Environment for R},
#>     author = {{RStudio Team}},
#>     organization = {RStudio, Inc.},
#>     address = {Boston, MA},
#>     year = {2016},
#>     url = {http://www.rstudio.com/},
#>   }
#> 
#> 
#> $mode
#> [1] "desktop"
#> 
#> $version
#> [1] ‘1.1.346’
#>