convertTheme {rstudioapi}	R Documentation

Convert a tmTheme to an RStudio Theme

Description

Converts a tmTheme to an rstheme and optionally adds and applies it to RStudio and returns the name of the theme.

Usage

convertTheme(themePath, add = TRUE, outputLocation = NULL,
  apply = FALSE, force = FALSE, globally = FALSE)

Arguments

themePath
A full or relative path to the tmTheme file to be converted.

add
Whether to add the newly converted theme to RStudio. Setting this to true will have the same impact as running { rstudioapi::convertTheme(<themePath>, outputLocation = <convertedThemePath>); rstudioapi::addTheme(<convertedThemePath>) }.
Default: TRUE.

outputLocation
A full or relative path where a copy of the converted theme will be saved. If this value is NULL, no copy will be saved.
Default: NULL.

apply
Whether to immediately apply the newly added theme. This paramater cannot be set to TRUE if add is set to FALSE. Setting this and add to TRUE has the same impact as running { rstudioapi::convertTheme(<themePath>, outputLocation = <convertedThemePath>); rstudioapi::addTheme(<convertedThemePath>); rstudioapi::applyTheme(<themeName>) }.
Default: FALSE.

force
Whether to force the operation and overwrite an existing file with the same name.
Default: FALSE.

globally
Whether to install this theme for the current user or all users. If set to TRUE this will attempt to install the theme for all users, which may require administrator privileges. Only applies when add is TRUE.
Default: FALSE.

Note

The convertTheme function was introduced in RStudio 1.2.879.

[Package rstudioapi version 0.10 Index]