dictionaries {rstudioapi}	R Documentation

Interact with RStudio's Dictionaries

Description

Interact with the hunspell dictionaries used by RStudio for spell checking.

Usage

dictionariesPath()

userDictionariesPath()

Details

dictionariesPath() gives a path to the dictionaries installed and distributed with RStudio.

userDictionariesPath() gives the path where users can provide their own custom hunspell dictionaries. See:

https://support.rstudio.com/hc/en-us/articles/200551916-Spelling-Dictionaries

for more information.

Note

The dictionariesPath() and userDictionariesPath() functions were introduced with RStudio 1.2.1202.

[Package rstudioapi version 0.10 Index]