executeCommand {rstudioapi}	R Documentation

Execute Command

Description

Executes an arbitrary RStudio command.

Usage

executeCommand(commandId, quiet = FALSE)

Arguments

commandId
The ID of the command to execute.

quiet
Whether to show an error if the command does not exist.

Details

Most menu commands and many buttons in RStudio can be invoked from the API using this method.

The quiet command governs the behavior of the function when the command does not exist. By default, an error is shown if you attempt to invoke a non-existent command. You should set this to FALSE when invoking a command that may not be available if you don't want your users to see an error.

The command is run asynchronously, so no status is returned.

Note

The executeCommand function was introduced in RStudio 1.2.1261.

[Package rstudioapi version 0.10 Index]