getDelegatedAzureToken {rstudioapi}	R Documentation

OAuth2 Tokens for Delegated Azure Resources

Description

When Workbench is using Azure Active Directory for sign-in, this function can return an OAuth2 token for a service Workbench users have delegated access to. This requires configuring delegated permissions in Azure itself.

Usage

getDelegatedAzureToken(resource)

Arguments

resource
The name of an Azure resource or service, normally a URL.

Examples

## Not run: 
getDelegatedAzureToken("https://storage.azure.com")

## End(Not run)

[Package rstudioapi version 0.15.0 Index]