getRStudioPackageDependencies {rstudioapi}	R Documentation

Get RStudio Package Dependencies

Description

Gets a list of the all the R packages that RStudio depends on in some way.

Usage

getRStudioPackageDependencies()

Details

The data frame of package dependencies contains the following columns:

name

The name of the R package.

version

The required minimum version of the R package.

location

Where RStudio expects the package to be, cran for a CRAN-like repository or embedded for development packages embedded in RStudio itself.

source

Whether the package should be installed from source.

Value

A data frame containing a row per R package.

Note

The getRStudioPackageDependencies function was introduced in RStudio 1.3.525.

[Package rstudioapi version 0.13 Index]