getThemeInfo {rstudioapi}	R Documentation

Retrieve Themes

Description

Retrieves a list with information about the current color theme used by RStudio.

Usage

getThemeInfo()

Details

A list is returned with the following elements:

editor

The name of the current editor theme, such as Textmate.

global

The name of the current global theme. One of Modern, Classic, or Sky.

dark

TRUE if the editor theme is dark, FALSE otherwise.

foreground

The current editor theme's default text foreground color, formatted as a CSS-compatible color string, such as rgb(1, 22, 39). Supported since RStudio 1.2.1214.

background

The current editor theme's default text background color, formatted as a CSS-compatible color string. Supported since RStudio 1.2.1214.

[Package rstudioapi version 0.15.0 Index]