launcherControlJob {rstudioapi}	R Documentation

Interact with (Control) a Job

Description

Interact with a job.

Usage

launcherControlJob(jobId, operation = c("suspend", "resume", "stop",
  "kill", "cancel"))

Arguments

jobId
The job id.

operation
The operation to execute. The operation should be one of c("suspend", "resume", "stop", "kill", "cancel"). Note that different launcher plugins support different subsets of these operations – consult your launcher plugin documentation to see which operations are supported.

[Package rstudioapi version 0.10 Index]