launcherGetJobs {rstudioapi}	R Documentation

Retrieve Workbench Job Information

Description

Retrieve information on Workbench jobs.

Usage

launcherGetJobs(
  statuses = NULL,
  fields = NULL,
  tags = NULL,
  includeSessions = FALSE
)

Arguments

statuses
Return only jobs whose status matches one of statuses. Valid statuses are: Pending, Running, Suspended, Failed, Finished, Killed, Canceled. When NULL, all jobs are returned.

fields
Return a subset of fields associated with each job object. When NULL, all fields associated with a particular job are returned.

tags
An optional set of tags. Only jobs that have been assigned one of these requested tags will be returned.

includeSessions
Boolean; include jobs which are also operating as RStudio R sessions?

See Also

Other job-launcher functionality: launcherAvailable(), launcherConfig(), launcherContainer(), launcherControlJob(), launcherGetInfo(), launcherGetJob(), launcherHostMount(), launcherNfsMount(), launcherPlacementConstraint(), launcherResourceLimit(), launcherSubmitJob(), launcherSubmitR()

[Package rstudioapi version 0.15.0 Index]