launcherPlacementConstraint {rstudioapi}	R Documentation

Define a Launcher Placement Constraint

Description

Define a launcher placement constraint, suitable for use with the placementConstraints argument to launcherSubmitJob().

Usage

launcherPlacementConstraint(name, value = NULL)

Arguments

name
The name of this placement constraint.

value
The value of the constraint. A job will only be placed on a requested node if the requested placement constraint is present.

See Also

Other job submission: launcherConfig, launcherContainer, launcherHostMount, launcherNfsMount, launcherResourceLimit, launcherSubmitJob, launcherSubmitR

[Package rstudioapi version 0.10 Index]