Submit a Launcher Job

Submit a launcher job. See https://docs.rstudio.com/job-launcher/latest/index.html for more information.

launcherSubmitJob(
  name,
  cluster = "Local",
  tags = NULL,
  command = NULL,
  exe = NULL,
  args = NULL,
  environment = NULL,
  stdin = NULL,
  stdoutFile = NULL,
  stderrFile = NULL,
  workingDirectory = NULL,
  host = NULL,
  container = NULL,
  exposedPorts = NULL,
  mounts = NULL,
  placementConstraints = NULL,
  resourceLimits = NULL,
  queues = NULL,
  config = NULL,
  user = Sys.getenv("USER"),
  applyConfigSettings = TRUE
)

Arguments

name
A descriptive name to assign to the job.

cluster
The name of the cluster this job should be submitted to.

tags
A set of user-defined tags, used for searching and querying jobs.

command
The command to run within the job. This is executed via the system shell. Only one of command or exe should be specified.

exe
The (fully pathed) executable to run within the job. Only one of command or exe should be specified.

args
An array of arguments to pass to the command / executable.

environment
A list of environment variables to be set for processes launched with this job.

stdin
Data to be written to stdin when the job process is launched.

stdoutFile
The file used for the job's generated standard output. Not all launcher plugins support this parameter.

stderrFile
The file used for the job's generated standard error. Not all launcher plugins support this parameter.

workingDirectory
The working directory to be used by the command / executable associated with this job.

host
The host that the job is running on, or the desired host during job submission.

container
The container to be used for launched jobs.

exposedPorts
The ports that are exposed by services running on a container. Only applicable to systems that support containers.

mounts
A list of mount points. See launcherHostMount() and launcherNfsMount() for more information.

placementConstraints
A list of placement constraints. See launcherPlacementConstraint() for more information.

resourceLimits
A list of resource limits. See launcherResourceLimit() for more information.

queues
A list of available submission queues for the cluster. Only applicable to batch systems like LSF.

config
A list of cluster-specific configuration options. See launcherConfig() for more information.

user
The user-name of the job owner.

applyConfigSettings
Apply server-configured mounts, exposedPorts, and environment, in addition to any specified in this call.

See also