launcherSubmitR {rstudioapi}	R Documentation

Execute an R Script as a Workbench Job

Description

Convenience function for running an R script as a Workbench job using whichever R is found on the path in the Workbench launcher cluster.

Usage


launcherSubmitR(script, cluster = "Local", container = NULL)

Arguments

script
Fully qualified path of R script. Must be a path that is available in the job container (if using containerized job cluster such as Kubernetes).

cluster
The name of the cluster this job should be submitted to.

container
The container to be used for launched jobs.

Details

See launcherSubmitJob() for running jobs with full control over command, environment, and so forth.

See Also

Other job-launcher functionality: launcherAvailable(), launcherConfig(), launcherContainer(), launcherControlJob(), launcherGetInfo(), launcherGetJob(), launcherGetJobs(), launcherHostMount(), launcherNfsMount(), launcherPlacementConstraint(), launcherResourceLimit(), launcherSubmitJob()

[Package rstudioapi version 0.16.0 Index]