persistent-values {rstudioapi}	R Documentation

Persistent Keys and Values

Description

Store persistent keys and values. Storage is per-project, if there is no project currently active then a global store is used.

Usage

setPersistentValue(name, value)

getPersistentValue(name)

Arguments

name
Key name

value
Key value

Value

The stored value as a character vector (NULL if no value of the specified name is available).

Note

The setPersistentValue and getPersistentValue functions were added in version 1.1.57 of RStudio.

[Package rstudioapi version 0.10 Index]