readPreference {rstudioapi}	R Documentation

Read Preference

Description

Reads a user interface preference, useful to remember preferences across different r sessions for the same user.

Usage

readPreference(name, default)

Arguments

name
The name of the preference.

default
The default value to use when the preference is not available.

Note

The readPreference function was added in version 1.1.67 of RStudio.

[Package rstudioapi version 0.10 Index]