sendToConsole {rstudioapi}	R Documentation

Send code to the R console

Description

Send code to the R console, and optionally execute it.

Usage

sendToConsole(code, execute = TRUE, echo = TRUE, focus = TRUE)

Arguments

code
The R code to be executed, as a character vector.

execute
Boolean; execute the code immediately or just enter the text into the console?

echo
Boolean; echo the code in the console as it is executed?

focus
Boolean; focus the console after sending code?

Note

The sendToConsole function was added in version 0.99.787 of RStudio.

Examples

## Not run: 
rstudioapi::sendToConsole(".Platform", execute = TRUE)

## End(Not run)

[Package rstudioapi version 0.13 Index]