sendToConsole {rstudioapi}	R Documentation

Send code to the R console

Description

Send code to the R console, and optionally execute it.

Usage

sendToConsole(code, execute = TRUE, echo = TRUE, focus = TRUE, animate = FALSE)

Arguments

code
The R code to be executed, as a character vector.

execute
Boolean; should the code be executed after being submitted to the console? If FALSE, code is submitted to the console but is not executed.

echo
Boolean; echo the code in the console as it is executed?

focus
Boolean; focus the console after sending code?

animate
Boolean; should the submitted code be animated, as if someone was typing it?

Note

The sendToConsole function was added in version 0.99.787 of RStudio.

Examples

## Not run: 
rstudioapi::sendToConsole(".Platform", execute = FALSE, animate = TRUE)

## End(Not run)

[Package rstudioapi version 0.15.0 Index]