showDialog {rstudioapi}	R Documentation

Show Dialog Box

Description

Shows a dialog box with a given title and contents.

Usage

showDialog(title, message, url = "", timeout = 60)

Arguments

title
The title to display in the dialog box.

message
A character vector with the contents to display in the main dialog area. Contents can contain the following HTML tags: "p", "em", "strong", "b" and "i".

url
An optional URL to display under the message.

timeout
A timeout (in seconds). When set, if the user takes longer than this timeout to provide a response, the request will be aborted.

Note

The showDialog function was added in version 1.1.67 of RStudio.

Examples

if (rstudioapi::isAvailable()) {
  rstudioapi::showDialog("Example Dialog", "This is an <b>example</b> dialog.")
}

[Package rstudioapi version 0.17.1 Index]