sourceMarkers {rstudioapi}	R Documentation

Display Source Markers

Description

Display user navigable source markers in a pane within RStudio

Usage

sourceMarkers(name, markers, basePath = NULL,
              autoSelect = c("none", "first", "error")) 

Arguments

name
Name of marker set (will replace any markers of the same name previously shown)

markers
List or data frame containing source markers (see below for details on how to specify markers)

basePath
Optional. If all source files are within a base path then specifying that path here will result in file names being displayed as relative paths. Note that in this case markers still need to specify source file names as full paths.

autoSelect
Optional. Automatically select and drive focus to either the first marker or the first marker that is an error.

Details

The markers argument can contains either a list of marker lists or a data frame with the appropriate marker columns. The fields in a marker are as follows (all are required):

type 	 Marker type ("error", "warning", "info", "style", or "usage")

file 	 Path to source file

line 	 Line number witin source file

column 	 Column number within line

message 	 Short descriptive message

Note that if the message field is of class "html" (i.e. inherits(message, "html") == TRUE) then it's contents will be treated as HTML.

Note

The sourceMarkers function was added in version 0.99.225 of RStudio.

[Package rstudioapi version 0.10 Index]