Activate Terminal

Ensure terminal is running and optionally bring to front in RStudio.

terminalActivate(id = NULL, show = TRUE)

Arguments

id
The terminal id. The id is obtained from terminalList(), terminalVisible(), terminalCreate(), or terminalExecute(). If NULL, the terminal tab will be selected but no specific terminal will be chosen.

show
If TRUE, bring the terminal to front in RStudio.

Note

The terminalActivate function was added in version 1.1.305 of RStudio.

Examples

# NOT RUN {
# create a hidden terminal and run a lengthy command
termId = rstudioapi::terminalCreate(show = FALSE)
rstudioapi::terminalSend(termId, "sleep 5\n")

# wait until a busy terminal is finished
while (rstudioapi::terminalBusy(termId)) {
  Sys.sleep(0.1)
}
print("Terminal available")#'

rstudioapi::terminalActivate(termId)
# }