terminalCreate {rstudioapi}	R Documentation

Create a Terminal

Description

Create a new Terminal.

Usage

terminalCreate(caption = NULL, show = TRUE, shellType = NULL)

Arguments

caption
The desired terminal caption. When NULL or blank, the terminal caption will be chosen by the system.

show
If FALSE, terminal won't be brought to front.

shellType
Shell type for the terminal: NULL or "default" to use the shell selected in Global Options. For Microsoft Windows, alternatives are "win-cmd" for 64-bit Command Prompt, "win-ps" for 64-bit PowerShell, "win-git-bash" for Git Bash, or "win-wsl-bash" for Bash on Windows Subsystem for Linux. On Linux, Mac, and RStudio Server "custom" will use the custom terminal defined in Global Options. If the requested shell type is not available, the default shell will be used, instead.

Value

The terminal identifier as a character vector (NULL if unable to create the terminal or the given terminal caption is already in use).

Note

The terminalCreate function was added in version 1.1.350 of RStudio and the ability to specify shellType was added in version 1.2.696.

Examples

## Not run: 
termId <- rstudioapi::terminalCreate('My Terminal')

## End(Not run)

[Package rstudioapi version 0.10 Index]