writeRStudioPreference {rstudioapi}	R Documentation

Write RStudio Preference

Description

Writes an internal RStudio IDE preference for the current user.

Usage

writeRStudioPreference(name, value)

Arguments

name
The name of the preference.

value
The value of the preference.

Details

RStudio IDE internal preferences include the values displayed in RStudio's Global Options dialog as well as a number of additional settings. Set them carefully; inappropriate values can cause unexpected behavior. See the RStudio Server Professional Administration Guide appendix for your version of RStudio for a full list of preference names and values.

Note

The writeRStudioPreference function was added in version 1.3.387 of RStudio.

See Also

writePreference, which can be used to store arbitrary user (non-RStudio) preferences.

readRStudioPreference, which reads internal RStudio IDE preferences.

Examples

## Not run: 
# Hide RStudio's toolbar.
rstudioapi::writeRStudioPreference("toolbar_visible", FALSE)

## End(Not run)

[Package rstudioapi version 0.13 Index]